[CSL-869] Move utxo utility functions into other modules, clarify `belongsTo'
