+++ src/Pos/Txp/Toil/Utxo.hs
@@ -1,9 +1,11 @@
 -- | Utxo functionality re-exports.
 
 module Pos.Txp.Toil.Utxo
-       ( module Pos.Txp.Toil.Utxo.Pure
-       , module Pos.Txp.Toil.Utxo.Functions
+       ( module Pos.Txp.Toil.Utxo.Functions
+       , module Pos.Txp.Toil.Utxo.Pure
+       , module Pos.Txp.Toil.Utxo.Util
        ) where
 
 import           Pos.Txp.Toil.Utxo.Functions
 import           Pos.Txp.Toil.Utxo.Pure
+import           Pos.Txp.Toil.Utxo.Util
